Parser module introduction
